REALLY STUPID BUG!!!! 2 fucking hours fixing this.
basically it's possible that a database object, or any other object could be created through the autoloader, but then perhaps it's equally possible that someone tries to include that file also using require() this would mean the autoloaded version wouldn't count and it would attempt to open again, cause the object to be redeclared and 99.999999999% of the time, all objects are autoloaded.  therefore sidestepping the problem.

but in this case, I was trying to create the <model> for the plugin and I would try three steps
1) import the model object from another plugin
2) require() the file
3) autoload the file

but since I mostly autoload (2) and (3) are in the wrong order, causing most typical occurances of trying to create a website wrong, so I debugged it for fucking ages, only to find that reversing the position of (2) and (3) would be ok!!

Signed-off-by: Christopher Thomas <chris.thomas@antimatter-studios.com>